Renamed AppChart.vue to Chart.vue
