+++ apps/explorer/src/modules/charts/components/Chart.vue
