+++ apps/explorer/src/modules/charts/components/history/ChartBlockSize.vue
@@ -1,5 +1,5 @@
 <template>
-  <app-chart
+  <chart
     type="line"
     :chart-title="title"
     :chart-description="description"
@@ -8,36 +8,46 @@
     :redraw="redraw"
     unfilled="true"
     @timeFrame="setTimeFrame"
-  >
-  </app-chart>
+  />
 </template>
 
 <script lang="ts">
-import AppChart from '@app/modules/charts/components/AppChart.vue'
+import Chart from '@app/modules/charts/components/Chart.vue'
 import { ChartMixin } from '@app/modules/charts/mixins'
 import { Component, Mixins } from 'vue-property-decorator'
 import { Events } from 'ethvm-common'
 
 @Component({
   components: {
-    AppChart
+    Chart
   }
 })
 export default class ChartBlockSize extends Mixins(ChartMixin) {
   /* Temp Event - Waiting for the event to be implemented, approximate release after 1.01 */
   newEvent = Events.getAverageDifficultyStats
 
-  // Lifecycle
+  /*
+  ===================================================================================
+    Lifecycle
+  ===================================================================================
+  */
+
   created() {
     this.setTitle(this.title)
     this.setLabel(this.labelString)
     this.setEvent(this.newEvent)
   }
 
-  // Computed
+  /*
+  ===================================================================================
+    Computed Values
+  ===================================================================================
+  */
+
   get title(): string {
     return this.$i18n.t('charts.avgBlockSize').toString()
   }
+
   get labelString(): string {
     return this.$i18n.t('charts.block-size-history').toString()
   }
