+++ apps/explorer/src/modules/charts/components/history/ChartGasLimit.vue
@@ -1,5 +1,5 @@
 <template>
-  <app-chart
+  <chart
     type="line"
     :chart-title="title"
     :chart-description="description"
@@ -12,27 +12,37 @@
 </template>
 
 <script lang="ts">
-import AppChart from '@app/modules/charts/components/AppChart.vue'
+import Chart from '@app/modules/charts/components/Chart.vue'
 import { ChartMixin } from '@app/modules/charts/mixins'
 import { Component, Mixins } from 'vue-property-decorator'
 import { Events } from 'ethvm-common'
 
 @Component({
   components: {
-    AppChart
+    Chart
   }
 })
 export default class ChartGasLimit extends Mixins(ChartMixin) {
   newEvent = Events.getAverageGasLimitStats
 
-  // Lifecycle
+  /*
+  ===================================================================================
+    Lifecycle
+  ===================================================================================
+  */
+
   created() {
     this.setTitle(this.title)
     this.setLabel(this.labelString)
     this.setEvent(this.newEvent)
   }
 
-  // Computed
+  /*
+  ===================================================================================
+    Computed Values
+  ===================================================================================
+  */
+
   get title() {
     return this.$i18n.t('charts.avgGasLimit').toString()
   }
