+++ apps/explorer/src/modules/charts/components/history/ChartTxSuccess.vue
@@ -1,5 +1,5 @@
 <template>
-  <app-chart
+  <chart
     type="line"
     :chart-title="title"
     :chart-description="description"
@@ -12,30 +12,41 @@
 </template>
 
 <script lang="ts">
-import AppChart from '@app/modules/charts/components/AppChart.vue'
+import Chart from '@app/modules/charts/components/Chart.vue'
 import { ChartMixin } from '@app/modules/charts/mixins'
 import { Component, Mixins } from 'vue-property-decorator'
 import { Events } from 'ethvm-common'
 
 @Component({
   components: {
-    AppChart
+    Chart
   }
 })
 export default class ChartTxSuccess extends Mixins(ChartMixin) {
   newEvent = Events.getSuccessfulTxStats
 
-  // Lifecycle
+  /*
+  ===================================================================================
+    Lifecycle
+  ===================================================================================
+  */
+
   created() {
     this.setTitle(this.title)
     this.setLabel(this.labelString)
     this.setEvent(this.newEvent)
   }
 
-  // Computed
+  /*
+  ===================================================================================
+    Computed Values
+  ===================================================================================
+  */
+
   get title(): string {
     return this.$i18n.t('charts.txSuccess').toString()
   }
+
   get labelString(): string {
     return this.$i18n.t('charts.tx-success-history').toString()
   }
