+++ apps/explorer/src/modules/charts/components/history/ChartUnclesNumber.vue
@@ -1,5 +1,5 @@
 <template>
-  <app-chart
+  <chart
     type="line"
     :chart-title="title"
     :chart-description="description"
@@ -12,14 +12,14 @@
 </template>
 
 <script lang="ts">
-import AppChart from '@app/modules/charts/components/AppChart.vue'
+import Chart from '@app/modules/charts/components/Chart.vue'
 import { ChartMixin } from '@app/modules/charts/mixins'
 import { Component, Mixins } from 'vue-property-decorator'
 import { Events } from 'ethvm-common'
 
 @Component({
   components: {
-    AppChart
+    Chart
   }
 })
 export default class ChartUnclesNumber extends Mixins(ChartMixin) {
@@ -27,17 +27,26 @@ export default class ChartUnclesNumber extends Mixins(ChartMixin) {
 
   newEvent = Events.getSuccessfulTxStats
 
-  // Lifecycle
+  /*
+  ===================================================================================
+    Lifecycle
+  ===================================================================================
+  */
   created() {
     this.setTitle(this.title)
     this.setLabel(this.labelString)
     this.setEvent(this.newEvent)
   }
 
-  // Computed
+  /*
+  ===================================================================================
+    Computed Values
+  ===================================================================================
+  */
   get title(): string {
     return this.$i18n.t('charts.totalUncles').toString()
   }
+
   get labelString(): string {
     return this.$i18n.t('charts.uncle-number-history').toString()
   }
